+++ packages/sdk/src/unsafe.ts
@@ -48,7 +48,6 @@ export {
 } from './lib/system/sudoTouchId.ts'
 export { installXcodeCli, isXcodeCliInstalled } from './lib/system/xcodeCli.ts'
 export { teardownGlobal } from './lib/teardown.ts'
-export { getDenvigVersion } from './lib/version.ts'
 
 export type { LaunchctlListItem } from './lib/services/launchctl.ts'
 export type { ServiceManagerProject } from './lib/services/manager.ts'